requirements.txt
jupyter
pandas
ploomber
pyarrow
scikit-learn
sklearn-evaluation
jupytext